2062: move helper function to application.ts, hide application card if application was already processed for applicant and verifier
