+++ administration/src/mui-modules/application-verification/ApplicationVerificationController.tsx
@@ -1,5 +1,4 @@
-import { Check, Close } from '@mui/icons-material'
-import { Alert, Box, Button, Card, Divider, Typography, styled } from '@mui/material'
+import { Stack } from '@mui/material'
 import { SnackbarProvider, useSnackbar } from 'notistack'
 import React, { ReactElement, useContext, useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
@@ -12,36 +11,17 @@ import {
   useVerifyOrRejectApplicationVerificationMutation,
 } from '../../generated/graphql'
 import { ProjectConfigContext } from '../../project-configs/ProjectConfigContext'
-import { parseApplication } from '../../shared/application'
-import JsonFieldView from '../../shared/components/JsonFieldView'
+import { applicationWasAlreadyProcessed, parseApplication } from '../../shared/application'
 import formatDateWithTimezone from '../../util/formatDate'
-import getApiBaseUrl from '../../util/getApiBaseUrl'
 import AlertBox from '../base/AlertBox'
 import getQueryResult from '../util/getQueryResult'
-import { applicationWasAlreadyProcessed } from './util'
-
-const ApplicationViewCard = styled(Card)`
-  max-width: 800px;
-  margin: 16px auto 16px auto;
-`
-
-const StyledAlert = styled(Alert)`
-  margin: 20px 0;
-  background-color: transparent;
-`
-
-const ButtonContainer = styled('div')`
-  display: flex;
-  width: inherit;
-  flex-direction: row;
-  justify-content: space-around;
-`
+import ApplicationVerifierView from './ApplicationVerifierView'
 
 type ApplicationVerificationProps = {
   applicationVerificationAccessKey: string
 }
 
-const ApplicationVerification = ({ applicationVerificationAccessKey }: ApplicationVerificationProps) => {
+const ApplicationVerificationController = ({ applicationVerificationAccessKey }: ApplicationVerificationProps) => {
   const { t } = useTranslation('applicationVerification')
   const [verificationFinished, setVerificationFinished] = useState(false)
   const config = useContext(ProjectConfigContext)
@@ -84,102 +64,51 @@ const ApplicationVerification = ({ applicationVerificationAccessKey }: Applicati
   const application = parseApplication(applicationQueryHandler.data.application)
 
   if (verification.rejectedDate || verification.verifiedDate) {
-    return <AlertBox severity='info' description={t('alreadyVerified')} />
+    return (
+      <Stack sx={{ flexGrow: 1, alignSelf: 'center', justifyContent: 'center', p: 2 }}>
+        <AlertBox severity='info' description={t('alreadyVerified')} />
+      </Stack>
+    )
   }
   if (application.status === ApplicationStatus.Withdrawn && application.statusResolvedDate) {
     return (
-      <AlertBox
-        severity='info'
-        description={t('withdrawMessageForVerifier', {
-          date: formatDateWithTimezone(application.statusResolvedDate, config.timezone),
-        })}
-      />
+      <Stack sx={{ flexGrow: 1, alignSelf: 'center', justifyContent: 'center', p: 2 }}>
+        <AlertBox
+          severity='info'
+          description={t('withdrawMessageForVerifier', {
+            date: formatDateWithTimezone(application.statusResolvedDate, config.timezone),
+          })}
+        />
+      </Stack>
     )
   }
   if (verificationFinished) {
     return <AlertBox title={t('verificationFinishedTitle')} description={t('verificationFinishedContent')} />
   }
 
-  const { createdDate: createdDateString, id } = application
-  const baseUrl = `${getApiBaseUrl()}/application/${config.projectId}/${id}`
-  const applicationVerificationDisabled = applicationWasAlreadyProcessed(application.status)
+  if (applicationWasAlreadyProcessed(application.status)) {
+    return (
+      <Stack sx={{ flexGrow: 1, alignSelf: 'center', justifyContent: 'center', p: 2 }}>
+        <AlertBox
+          title={t('applicationAlreadyProcessed')}
+          description={<Trans i18nKey='applicationVerification:applicationAlreadyProcessedHint' />}
+          severity='info'
+          borderless
+        />
+      </Stack>
+    )
+  }
 
   return (
-    <Box sx={{ flexGrow: 1, overflow: 'auto', alignItems: 'safe center' }}>
-      <ApplicationViewCard elevation={2}>
-        <div style={{ overflow: 'visible', padding: '20px' }}>
-          <Typography sx={{ mb: '12px' }} variant='h4'>
-            {config.name}
-          </Typography>
-          <Typography sx={{ my: '8px' }} variant='body1'>
-            {t('greeting', { contactName: verification.contactName })}
-            <br />
-            <br />
-            <Trans
-              i18nKey='applicationVerification:text'
-              values={{ organizationName: verification.organizationName }}
-            />
-          </Typography>
-          <Divider style={{ margin: '24px 0px' }} />
-          <Typography variant='h6' sx={{ mb: '8px' }}>
-            Antrag vom {formatDateWithTimezone(createdDateString, config.timezone)}
-          </Typography>
-          <JsonFieldView
-            jsonField={application.jsonValue}
-            baseUrl={baseUrl}
-            hierarchyIndex={0}
-            attachmentAccessible={false}
-            expandedRoot
-          />
-          <Divider style={{ margin: '24px 0px' }} />
-          {applicationVerificationDisabled ? (
-            <Box sx={{ marginBottom: 3 }}>
-              <AlertBox
-                title={t('applicationAlreadyProcessed')}
-                description={t('applicationAlreadyProcessedHint')}
-                severity='info'
-                borderless
-              />
-            </Box>
-          ) : (
-            <>
-              <Typography sx={{ mt: '8px' }} variant='body1'>
-                <Trans
-                  i18nKey='applicationVerification:confirmationMessage'
-                  values={{ organizationName: verification.organizationName }}
-                />
-              </Typography>
-              <StyledAlert severity='warning'>
-                <Trans i18nKey='applicationVerification:confirmationNote' />
-              </StyledAlert>
-            </>
-          )}
-
-          <ButtonContainer>
-            <Button
-              variant='contained'
-              color='error'
-              disabled={applicationVerificationDisabled}
-              endIcon={<Close />}
-              onClick={() => submitApplicationVerification(false)}>
-              {t('rejectButton')}
-            </Button>
-            <Button
-              variant='contained'
-              color='success'
-              disabled={applicationVerificationDisabled}
-              endIcon={<Check />}
-              onClick={() => submitApplicationVerification(true)}>
-              {t('confirmationButton')}
-            </Button>
-          </ButtonContainer>
-        </div>
-      </ApplicationViewCard>
-    </Box>
+    <ApplicationVerifierView
+      verification={verification}
+      application={application}
+      submitApplicationVerification={submitApplicationVerification}
+    />
   )
 }
 
-const ApplicationVerificationController = (): ReactElement => {
+const ControllerWithAccessKey = (): ReactElement => {
   const { t } = useTranslation('applicationVerification')
   const { applicationVerificationAccessKey } = useParams()
 
@@ -195,9 +124,9 @@ const ApplicationVerificationController = (): ReactElement => {
 
   return (
     <SnackbarProvider>
-      <ApplicationVerification applicationVerificationAccessKey={applicationVerificationAccessKey} />
+      <ApplicationVerificationController applicationVerificationAccessKey={applicationVerificationAccessKey} />
     </SnackbarProvider>
   )
 }
 
-export default ApplicationVerificationController
+export default ControllerWithAccessKey
