--- administration/src/mui-modules/application-verification/util/index.ts
@@ -1,6 +1,0 @@
-import { ApplicationStatus } from '../../../generated/graphql'
-
-export const applicationWasAlreadyProcessed = (status: ApplicationStatus): boolean =>
-  status === ApplicationStatus.Approved ||
-  status === ApplicationStatus.ApprovedCardCreated ||
-  status === ApplicationStatus.Rejected
